Add files via upload
Adding lines of code containing main execution of code at the bottom of the file. This code calls the user input functions to obtain input coordinates from user as well as clearance and radius and then call Djikstra and bactracking functions to obtain shortest path. Then, the pygame animation is called to simulate the Djikstra algorithm as it works through the map and obstacles space until it reaches the desired goal node.